add monaco editor with example
